move typing result and error types to its own module
