+++ src/ty/mod.rs
@@ -3,8 +3,10 @@ use std::fmt;
 use crate::ast::Name;
 
 mod ty_check;
+mod result;
 
-pub use ty_check::{expect_ty, ty_check, TyError, TyResult};
+pub use ty_check::{expect_ty, ty_check};
+pub use result::{TyError, TyResult};
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum Ty {
